// src/modules/secretary/components/Patients/CreateUserModal/interfaces.ts
export interface ICreateUserModal {
  close(): void;
}

export interface IFormData {
  name: string;
  email: string;
  phone: string;
  register: string;
  birthday: string;
  medicalRecord: string;
  gender: string;
  color: string;
  naturalness: string;
  maritalStatus: string;
  cpf: string;
  cpfOwner: string;
  educationLevel: string;
  profession: string;
  healthInsurance: string;
  registration: string;
  cep: string;
  street: string;
  complement: string;
  number: string;
  neighborhood: string;
  city: string;
  state: string;
}